consolidate probe metadata tests into one module
