--- tests/probes/test_probe_docs.py
@@ -1,27 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-
-import importlib
-import pytest
-import re
-
-from garak import _plugins
-
-PROBES = [classname for (classname, active) in _plugins.enumerate_plugins("probes")]
-
-
-@pytest.mark.parametrize("plugin_name", PROBES)
-def test_check_docstring(plugin_name):
-    plugin_name_parts = plugin_name.split(".")
-    module_name = "garak." + ".".join(plugin_name_parts[:-1])
-    class_name = plugin_name_parts[-1]
-    mod = importlib.import_module(module_name)
-    doc = getattr(getattr(mod, class_name), "__doc__")
-    doc_paras = re.split(r"\s*\n\s*\n\s*", doc)
-    assert (
-        len(doc_paras) >= 2
-    )  # probe class doc should have a summary, two newlines, then a paragraph giving more depth, then optionally more words
-    assert (
-        len(doc_paras[0]) > 0
-    )  # the first paragraph of the probe docstring should not be empty
