--- tests/probes/test_probe_tags.py
@@ -1,35 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-import importlib
-import pytest
-import re
-
-from garak import _plugins, _config
-
-PROBES = [classname for (classname, active) in _plugins.enumerate_plugins("probes")]
-
-with open(
-    _config.transient.package_dir / "resources" / "misp_descriptions.tsv",
-    "r",
-    encoding="utf-8",
-) as misp_data:
-    MISP_TAGS = [line.split("\t")[0] for line in misp_data.read().split("\n")]
-
-
-@pytest.mark.parametrize("classname", PROBES)
-def test_tag_format(classname):
-    plugin_name_parts = classname.split(".")
-    module_name = "garak." + ".".join(plugin_name_parts[:-1])
-    class_name = plugin_name_parts[-1]
-    mod = importlib.import_module(module_name)
-    cls = getattr(mod, class_name)
-    assert (
-        cls.tags != [] or cls.active == False
-    )  # all probes should have at least one tag
-    for tag in cls.tags:  # should be MISP format
-        assert type(tag) == str
-        for part in tag.split(":"):
-            assert re.match(r"^[A-Za-z0-9_\-]+$", part)
-        if tag.split(":")[0] != "payload":
-            assert tag in MISP_TAGS
