+++ tests/probes/test_probes.py
@@ -1,10 +1,14 @@
+# SPDX-FileCopyrightText: Copyright (c) 2023 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
 import importlib
 import pytest
 import re
 
-from garak import _plugins
+from garak import _config, _plugins
 
 PROBES = [classname for (classname, active) in _plugins.enumerate_plugins("probes")]
+
 DETECTORS = [
     classname
     for (classname, active) in _plugins.enumerate_plugins(
@@ -12,6 +16,15 @@ DETECTORS = [
     )
 ]
 DETECTOR_BARE_NAMES = [".".join(d.split(".")[1:]) for d in DETECTORS]
+
+BCP_LENIENT_RE = re.compile(r"[a-z]{2}([\-A-Za-z]*)")
+
+with open(
+    _config.transient.package_dir / "resources" / "misp_descriptions.tsv",
+    "r",
+    encoding="utf-8",
+) as misp_data:
+    MISP_TAGS = [line.split("\t")[0] for line in misp_data.read().split("\n")]
 
 
 @pytest.mark.parametrize("classname", PROBES)
@@ -56,9 +69,6 @@ def test_probe_structure(classname):
     assert unsupported_defaults == []
 
 
-bcp_lenient_re = re.compile(r"[a-z]{2}([\-A-Za-z]*)")
-
-
 @pytest.mark.parametrize("classname", PROBES)
 def test_probe_metadata(classname):
     p = _plugins.load_plugin(classname)
@@ -70,7 +80,7 @@ def test_probe_metadata(classname):
     bcp47_parts = p.bcp47.split(",")
     for bcp47_part in bcp47_parts:
         assert bcp47_part == "*" or re.match(
-            bcp_lenient_re, bcp47_part
+            BCP_LENIENT_RE, bcp47_part
         ), "langs must be described with either * or a bcp47 code"
     assert isinstance(
         p.doc_uri, str
@@ -82,3 +92,37 @@ def test_probe_metadata(classname):
     assert isinstance(p.modality, dict), "probes need to describe available modalities"
     assert "in" in p.modality, "probe modalities need an in descriptor"
     assert isinstance(p.modality["in"], set), "modality descriptors must be sets"
+
+
+@pytest.mark.parametrize("plugin_name", PROBES)
+def test_check_docstring(plugin_name):
+    plugin_name_parts = plugin_name.split(".")
+    module_name = "garak." + ".".join(plugin_name_parts[:-1])
+    class_name = plugin_name_parts[-1]
+    mod = importlib.import_module(module_name)
+    doc = getattr(getattr(mod, class_name), "__doc__")
+    doc_paras = re.split(r"\s*\n\s*\n\s*", doc)
+    assert (
+        len(doc_paras) >= 2
+    )  # probe class doc should have a summary, two newlines, then a paragraph giving more depth, then optionally more words
+    assert (
+        len(doc_paras[0]) > 0
+    )  # the first paragraph of the probe docstring should not be empty
+
+
+@pytest.mark.parametrize("classname", PROBES)
+def test_tag_format(classname):
+    plugin_name_parts = classname.split(".")
+    module_name = "garak." + ".".join(plugin_name_parts[:-1])
+    class_name = plugin_name_parts[-1]
+    mod = importlib.import_module(module_name)
+    cls = getattr(mod, class_name)
+    assert (
+        cls.tags != [] or cls.active == False
+    )  # all probes should have at least one tag
+    for tag in cls.tags:  # should be MISP format
+        assert type(tag) == str
+        for part in tag.split(":"):
+            assert re.match(r"^[A-Za-z0-9_\-]+$", part)
+        if tag.split(":")[0] != "payload":
+            assert tag in MISP_TAGS
